+++ src/types/inspect.ts
@@ -1,0 +1,6 @@
+import type { Commitment, createSolanaRpc } from "@solana/web3.js";
+
+export type InspectorBaseArgs = {
+  rpc: ReturnType<typeof createSolanaRpc>;
+  commitment?: Commitment;
+};
